requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
google-genai==0.8.0
google-auth==2.35.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
azure-storage-blob==12.19.1
psycopg2-binary==2.9.9
mammoth==1.6.0
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.32.3
